refactor: clean up lib.rs
