+++ src/ecc/mod.rs
@@ -1,2 +1,3 @@
 pub mod ecc_encryptable;
 pub mod ecc_signable;
+pub mod keypair;
