+++ src/lib.rs
@@ -1,16 +1,28 @@
 #![warn(clippy::just_underscores_and_digits)]
-use capy_kem::constants::parameter_sets::KEM_768;
-use capy_kem::fips203::keygen::k_pke_keygen;
-
-use rand::{thread_rng, RngCore};
+use ecc::ecc_signable::Signature;
 use serde::{Deserialize, Serialize};
-use sha3::{
-    aux_functions::byte_utils::{bytes_to_scalar, get_date_and_time_as_string},
-    shake_functions::kmac_xof,
-};
 use std::fs::File;
 use std::io::Read;
-use tiny_ed448_goldilocks::curve::{extended_edwards::ExtendedPoint, field::scalar::Scalar};
+use tiny_ed448_goldilocks::curve::extended_edwards::ExtendedPoint;
+
+const RATE_IN_BYTES: usize = 136; // SHA3-256 r = 1088 / 8 = 136
+
+#[cfg(test)]
+const NIST_DATA_SPONGE_INIT: [u8; 200] = [
+    0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
+    0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
+    0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f,
+    0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x3a, 0x3b, 0x3c, 0x3d, 0x3e, 0x3f,
+    0x40, 0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47, 0x48, 0x49, 0x4a, 0x4b, 0x4c, 0x4d, 0x4e, 0x4f,
+    0x50, 0x51, 0x52, 0x53, 0x54, 0x55, 0x56, 0x57, 0x58, 0x59, 0x5a, 0x5b, 0x5c, 0x5d, 0x5e, 0x5f,
+    0x60, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66, 0x67, 0x68, 0x69, 0x6a, 0x6b, 0x6c, 0x6d, 0x6e, 0x6f,
+    0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77, 0x78, 0x79, 0x7a, 0x7b, 0x7c, 0x7d, 0x7e, 0x7f,
+    0x80, 0x81, 0x82, 0x83, 0x84, 0x85, 0x86, 0x87, 0x88, 0x89, 0x8a, 0x8b, 0x8c, 0x8d, 0x8e, 0x8f,
+    0x90, 0x91, 0x92, 0x93, 0x94, 0x95, 0x96, 0x97, 0x98, 0x99, 0x9a, 0x9b, 0x9c, 0x9d, 0x9e, 0x9f,
+    0xa0, 0xa1, 0xa2, 0xa3, 0xa4, 0xa5, 0xa6, 0xa7, 0xa8, 0xa9, 0xaa, 0xab, 0xac, 0xad, 0xae, 0xaf,
+    0xb0, 0xb1, 0xb2, 0xb3, 0xb4, 0xb5, 0xb6, 0xb7, 0xb8, 0xb9, 0xba, 0xbb, 0xbc, 0xbd, 0xbe, 0xbf,
+    0xc0, 0xc1, 0xc2, 0xc3, 0xc4, 0xc5, 0xc6, 0xc7,
+];
 
 /// A simple error type
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
@@ -48,162 +60,17 @@ pub mod sha3 {
     /// Submodule that implements the Keccak-f[1600] permutation
     pub mod keccakf;
 
-    pub mod sha3_encryptable;
     pub mod sha3_hashable;
     pub mod shake_functions;
     /// Submodule that implements the sponge construction
     pub mod sponge;
+    pub mod sponge_encryptable;
 }
 
 pub mod aes {
     pub mod aes_constants;
     pub mod aes_encryptable;
     pub mod aes_functions;
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct KEMKey {
-    pub ek: Vec<u8>,
-    pub dk: Vec<u8>,
-    pub rand_bytes: [u8; 32],
-}
-
-impl KEMKey {
-    pub fn kem_keygen() -> KEMKey {
-        let mut rng = thread_rng();
-        let mut rand_bytes = [0u8; 32];
-
-        // generate randomness for the KEM
-        rng.fill_bytes(&mut rand_bytes);
-        let (ek, dk) = k_pke_keygen::<KEM_768>(&rand_bytes);
-
-        KEMKey { ek, dk, rand_bytes }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-/// An object containing the necessary fields for Schnorr signatures.
-pub struct Signature {
-    /// keyed hash of signed message
-    pub h: Vec<u8>,
-    /// public nonce
-    pub z: Scalar,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-/// An object containing the fields necessary to represent an asymmetric keypair.
-pub struct KeyPair {
-    /// String indicating the owner of the key, can be arbitrary
-    pub owner: String,
-    /// Public encryption key
-    pub pub_key: ExtendedPoint,
-    /// value representing secret scalar, None if KeyType is PUBLIC
-    pub priv_key: Vec<u8>,
-    /// Date key was generated
-    pub date_created: String,
-}
-
-impl KeyPair {
-    /// # Asymmetric [`KeyPair`] Generation
-    /// Generates a (Schnorr/ECDHIES) key pair from passphrase pw.
-    ///
-    /// ## Algorithm:
-    /// * s ← kmac_xof(pw, “”, 448, “K”); s ← 4s
-    /// * 𝑉 ← s*𝑮
-    /// * key pair: (s, 𝑉)
-    /// ## Arguments:
-    /// * pw: &Vec<u8> : password as bytes, can be blank but shouldnt be
-    /// * owner: String : A label to indicate the owner of the key
-    /// * curve: [`EdCurves`] : The selected Edwards curve
-    /// ## Returns:
-    /// * return  -> [`KeyPair`]: Key object containing owner, private key, public key x and y coordinates, and timestamp.
-    /// verification key 𝑉 is hashed together with the message 𝑚
-    /// and the nonce 𝑈: hash (𝑚, 𝑈, 𝑉) .
-    /// ## Usage:
-    /// ```
-    /// use capycrypt::{
-    ///     KeyEncryptable,
-    ///     KeyPair,
-    ///     Message,
-    ///     sha3::aux_functions::byte_utils::get_random_bytes,
-    ///     SecParam,
-    /// };
-    ///
-    /// // Get 5mb random data
-    /// let mut msg = Message::new(get_random_bytes(5242880));
-    /// // Create a new private/public keypair
-    /// let key_pair = KeyPair::new(&get_random_bytes(32), "test key".to_string(), &SecParam::D512).expect("Failed to create key pair");
-    ///
-    /// // Encrypt the message
-    /// msg.key_encrypt(&key_pair.pub_key, &SecParam::D512);
-    //  Decrypt the message
-    /// msg.key_decrypt(&key_pair.priv_key);
-    /// // Verify successful operation
-    /// msg.op_result.expect("Asymmetric decryption failed");    
-    /// ```
-    #[allow(non_snake_case)]
-    pub fn new(pw: &[u8], owner: String, d: &SecParam) -> Result<KeyPair, OperationError> {
-        let data = kmac_xof(pw, &[], 448, "SK", d)?;
-        let s: Scalar = bytes_to_scalar(data).mul_mod(&Scalar::from(4_u64));
-        let V = ExtendedPoint::generator() * s;
-        Ok(KeyPair {
-            owner,
-            pub_key: V,
-            priv_key: pw.to_vec(),
-            date_created: get_date_and_time_as_string(),
-        })
-    }
-
-    /// # KeyPair Saving
-    ///
-    /// Saves the key pair to a JSON file.
-    ///
-    /// ## Usage:
-    ///
-    /// ```rust
-    /// use capycrypt::KeyPair;
-    /// use capycrypt::SecParam;
-    ///
-    /// let key_pair = KeyPair::new("password".as_bytes(), "owner".to_string(), &SecParam::D512)
-    ///     .expect("Failed to create key pair");
-    ///
-    /// // key_pair.write_to_file("keypai1r.json").expect("Failed to save key pair");
-    pub fn write_to_file(&self, filename: &str) -> std::io::Result<()> {
-        let json_key_pair = serde_json::to_string_pretty(self).unwrap();
-        std::fs::write(filename, json_key_pair)
-    }
-
-    /// # KeyPair Loading
-    ///
-    /// Reads a JSON file and creates a `KeyPair` from its contents.
-    ///
-    /// ## Errors:
-    ///
-    /// Returns an error if:
-    /// - The file cannot be opened or read.
-    /// - The JSON content cannot be parsed into a `KeyPair`.
-    ///
-    /// ## Usage:
-    ///
-    /// ```rust
-    /// use capycrypt::KeyPair;
-    ///
-    /// // Assuming "keypair.json" contains a serialized KeyPair
-    /// match KeyPair::read_from_file("keypair.json") {
-    ///     Ok(key_pair) => {
-    ///         println!("Loaded KeyPair: {:?}", key_pair);
-    ///     },
-    ///     Err(err) => eprintln!("Error loading KeyPair: {}", err),
-    /// }
-    /// ```
-    pub fn read_from_file(filename: &str) -> Result<KeyPair, Box<dyn std::error::Error>> {
-        let mut file = File::open(filename)?;
-        let mut contents = String::new();
-        file.read_to_string(&mut contents)?;
-
-        let keypair: KeyPair = serde_json::from_str(&contents)?;
-        Ok(keypair)
-    }
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
@@ -392,54 +259,3 @@ impl BitLength for OutputLength {
         self.value()
     }
 }
-
-pub trait AesEncryptable {
-    fn aes_encrypt_cbc(&mut self, key: &[u8]) -> Result<(), OperationError>;
-    fn aes_decrypt_cbc(&mut self, key: &[u8]) -> Result<(), OperationError>;
-    fn aes_encrypt_ctr(&mut self, key: &[u8]) -> Result<(), OperationError>;
-    fn aes_decrypt_ctr(&mut self, key: &[u8]) -> Result<(), OperationError>;
-}
-
-pub trait Hashable {
-    fn compute_hash_sha3(&mut self, d: &SecParam) -> Result<(), OperationError>;
-    fn compute_tagged_hash(&mut self, pw: &[u8], s: &str, d: &SecParam);
-}
-
-pub trait SpongeEncryptable {
-    fn sha3_encrypt(&mut self, pw: &[u8], d: &SecParam) -> Result<(), OperationError>;
-    fn sha3_decrypt(&mut self, pw: &[u8]) -> Result<(), OperationError>;
-}
-
-pub trait KeyEncryptable {
-    fn key_encrypt(&mut self, pub_key: &ExtendedPoint, d: &SecParam) -> Result<(), OperationError>;
-    fn key_decrypt(&mut self, pw: &[u8]) -> Result<(), OperationError>;
-}
-
-pub trait KEMEncryptable {
-    fn kem_encrypt(&mut self, key: &KEMKey, d: &SecParam) -> Result<(), OperationError>;
-    fn kem_decrypt(&mut self, key: &KEMKey) -> Result<(), OperationError>;
-}
-
-pub trait Signable {
-    fn sign(&mut self, key: &KeyPair, d: &SecParam) -> Result<(), OperationError>;
-    fn verify(&mut self, pub_key: &ExtendedPoint) -> Result<(), OperationError>;
-}
-
-const RATE_IN_BYTES: usize = 136; // SHA3-256 r = 1088 / 8 = 136
-
-#[cfg(test)]
-const NIST_DATA_SPONGE_INIT: [u8; 200] = [
-    0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
-    0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
-    0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f,
-    0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x3a, 0x3b, 0x3c, 0x3d, 0x3e, 0x3f,
-    0x40, 0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47, 0x48, 0x49, 0x4a, 0x4b, 0x4c, 0x4d, 0x4e, 0x4f,
-    0x50, 0x51, 0x52, 0x53, 0x54, 0x55, 0x56, 0x57, 0x58, 0x59, 0x5a, 0x5b, 0x5c, 0x5d, 0x5e, 0x5f,
-    0x60, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66, 0x67, 0x68, 0x69, 0x6a, 0x6b, 0x6c, 0x6d, 0x6e, 0x6f,
-    0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77, 0x78, 0x79, 0x7a, 0x7b, 0x7c, 0x7d, 0x7e, 0x7f,
-    0x80, 0x81, 0x82, 0x83, 0x84, 0x85, 0x86, 0x87, 0x88, 0x89, 0x8a, 0x8b, 0x8c, 0x8d, 0x8e, 0x8f,
-    0x90, 0x91, 0x92, 0x93, 0x94, 0x95, 0x96, 0x97, 0x98, 0x99, 0x9a, 0x9b, 0x9c, 0x9d, 0x9e, 0x9f,
-    0xa0, 0xa1, 0xa2, 0xa3, 0xa4, 0xa5, 0xa6, 0xa7, 0xa8, 0xa9, 0xaa, 0xab, 0xac, 0xad, 0xae, 0xaf,
-    0xb0, 0xb1, 0xb2, 0xb3, 0xb4, 0xb5, 0xb6, 0xb7, 0xb8, 0xb9, 0xba, 0xbb, 0xbc, 0xbd, 0xbe, 0xbf,
-    0xc0, 0xc1, 0xc2, 0xc3, 0xc4, 0xc5, 0xc6, 0xc7,
-];
