+++ src/sha3/sponge_encryptable.rs
@@ -1,8 +1,13 @@
 use super::shake_functions::kmac_xof;
 use crate::{
     sha3::aux_functions::byte_utils::{get_random_bytes, xor_bytes},
-    Message, OperationError, SecParam, SpongeEncryptable,
+    Message, OperationError, SecParam,
 };
+
+pub trait SpongeEncryptable {
+    fn sha3_encrypt(&mut self, pw: &[u8], d: &SecParam) -> Result<(), OperationError>;
+    fn sha3_decrypt(&mut self, pw: &[u8]) -> Result<(), OperationError>;
+}
 
 impl SpongeEncryptable for Message {
     /// # Symmetric Encryption
